documentation
The principal goal guiding the design of any encryption algorithm must
be security against unauthorized attacks. Within the last decade, there
has been a vast increase in  the accumulation and communication of
digital computer data in both the private and public sectors. Much of
this information has a significant value, either directly or indirectly,
which requires protection. The algorithms uniquely define the
mathematical steps required to transform data into a cryptographic
cipher and also to transform the cipher back to the original form.
Performance and security level is the main characteristics that
differentiate one encryption algorithm from another. Here introduces a
new method to enhance the performance of the Data Encryption Standard
(DES) algorithm is introduced here. This is done by replacing the
predefined XOR operation applied during the 16 round of the standard
algorithm by a new operation depends on using two keys, each key
consists of a combination of 4 states (0, 1, 2, 3) instead of the
ordinary 2 state key (0, 1). This replacement adds a new level of
protection strength and more robustness against breaking methods.